% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_text_data.R
\name{label_text_data}
\alias{label_text_data}
\title{label_text_data}
\usage{
label_text_data(
  data_frame_name,
  labels,
  app_name = "Text Classification App",
  num_rows = 10,
  user_name = "user"
)
}
\arguments{
\item{data_frame_name}{A string specifying the name of the data frame in the global environment that contains the text to be labeled. This data frame should exist in the global environment prior to calling this function.}

\item{labels}{A character vector containing the possible labels for annotation. It should have at least two elements. This parameter defines the options available to the user for labeling the text data.}

\item{app_name}{Optional; a string that sets the title of the annotation application. Default is "Text Classification App". This name appears in the title panel of the Shiny app.}

\item{num_rows}{Optional; an integer indicating the number of rows to be randomly sampled from the data frame for labeling in a single session. Default is 10. This determines the workload for a single annotation session.}

\item{user_name}{Optional; a string representing the name of the user. Default is "user". This is used for personalizing the session and naming the output file containing the labeled data.}
}
\description{
This function initiates an interactive text annotation application using Shiny. It's designed to facilitate the labeling of text data for classification tasks. Users can annotate a specified number of text entries from a given data frame with predefined labels.
}
\examples{
if (interactive()) {
  df <- data.frame(text = c("Sample text 1", "Sample text 2"), stringsAsFactors = FALSE)
  label_text_data("df", c("Positive", "Negative"), "My Annotation App", 5, "user1")
}
}
